requirements.txt
discord==2.3.2
discord.py==2.4.0
openai==1.61.1
python-dotenv==1.0.1
pytest-cov==6.0.0
pytest==8.3.4
pytest-asyncio==0.25.3
